User logout and fixing post deletion

+++ app/Http/Controllers/PostController.php
@@ -5,13 +5,14 @@ namespace App\Http\Controllers;
 use App\Http\Controllers\Controller;
 
 use App\Post;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 
 class PostController extends Controller {
 
   public function getDashboard()
   {
-    $posts = Post::all();
+    $posts = Post::orderBy('created_at', 'desc')->get();
     return view('dashboard', ['posts' => $posts]);
   }
 
@@ -33,6 +34,12 @@ class PostController extends Controller {
 
   public function getDeletePost($post_id) {
     $post = Post::where('id', $post_id)->first();
+
+    // protection to do not allow other user deleting this posts
+    if(Auth::user() != $post->user) {
+      return redirect()->back();
+    }
+
     $post->delete();
     return redirect()->route('dashboard', ['message' => 'Successfully delete!']);
   }
